+++ arch/mips/math-emu/sp_fdp.c
@@ -43,7 +43,7 @@ ieee754sp ieee754sp_fdp(ieee754dp x)
 		SETCX(IEEE754_INVALID_OPERATION);
 		return ieee754sp_nanxcpt(ieee754sp_indef(), "fdp");
 	case IEEE754_CLASS_QNAN:
-		nan = buildsp(xs, SP_EMAX + 1 + SP_EBIAS, (unsigned long)
+		nan = buildsp(xs, SP_EMAX + 1 + SP_EBIAS, (u32)
 				(xm >> (DP_MBITS - SP_MBITS)));
 		if (!ieee754sp_isnan(nan))
 			nan = ieee754sp_indef();
@@ -57,7 +57,7 @@ ieee754sp ieee754sp_fdp(ieee754dp x)
 		SETCX(IEEE754_UNDERFLOW);
 		SETCX(IEEE754_INEXACT);
 		if ((ieee754_csr.rm == IEEE754_RU && !xs) ||
-		    (ieee754_csr.rm == IEEE754_RD && xs))
+				(ieee754_csr.rm == IEEE754_RD && xs))
 			return ieee754sp_xcpt(ieee754sp_mind(xs), "fdp", x);
 		return ieee754sp_xcpt(ieee754sp_zero(xs), "fdp", x);
 	case IEEE754_CLASS_NORM:
@@ -65,7 +65,7 @@ ieee754sp ieee754sp_fdp(ieee754dp x)
 	}
 
 	{
-		unsigned long rm;
+		u32 rm;
 
 		/* convert from DP_MBITS to SP_MBITS+3 with sticky right shift 
 		 */
